fix: Separate components from onboarding index
